fix: remove duplicate previous_hash field definition (#2662)
